Changes, Plus added Save Load Feature
Controller has a String in it called NEAT_SAVE_LOAD_FILE
set this to whatever file name with full path you want to save/load
to/from. Use a ".sav" at end of file. Remember to change this because
current default value will not be friendly to your personal computer